cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "waifuos-cli"
version = "0.0.0"
description = "CLI for WaifuOS: The software suite for creating your waifu and the world where your waifu lives.💕"
authors = [{ name = "uezo", email = "uezo@uezo.net" }]
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "httpx>=0.27.2",
    "pillow>=11.1.0",
]
[project.scripts]
waifu = "waifuos_cli.script:main"
[tool.setuptools]
package-dir = {"" = "src"}
[tool.setuptools.packages.find]
where = ["src"]
